pyproject.toml
[project]
name = "dystream"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "aiofiles>=23.2.1",
    "audioread==3.1.0",
    "diffusers>=0.24.0",
    "einops>=0.7.0",
    "fastapi>=0.104.0",
    "fire==0.7.1",
    "gitpython==3.1.46",
    "gradio==6.5.1",
    "huggingface-hub==0.36.0",
    "imageio==2.37.2",
    "imageio-ffmpeg==0.6.0",
    "librosa==0.11.0",
    "lightning==2.6.0",
    "lpips==0.1.4",
    "mediapipe==0.10.21",
    "moviepy==1.0.3",
    "numpy>=1.24.0",
    "omegaconf>=2.3.0",
    "opencv-contrib-python>=4.11.0.86",
    "opencv-python>=4.8.0",
    "pandas>=2.3.3",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "python-dotenv==1.2.1",
    "python-multipart>=0.0.6",
    "pytorch-lightning>=2.1.0",
    "pyyaml>=6.0",
    "regex==2026.1.15",
    "requests==2.32.5",
    "safetensors==0.7.0",
    "scikit-image==0.26.0",
    "scikit-learn==1.8.0",
    "scipy>=1.11.0",
    "sentencepiece==0.2.1",
    "sounddevice==0.5.5",
    "soundfile==0.13.1",
    "soxr==1.0.0",
    "timm==1.0.24",
    "tokenizers==0.22.2",
    "torch>=2.0.0",
    "torch-ema",
    "torchmetrics==1.8.2",
    "torchvision>=0.15.0",
    "tqdm>=4.66.0",
    "transformers>=4.35.0",
    "uvicorn[standard]>=0.24.0",
    "wandb>=0.16.0",
    "websockets>=12.0",
    "pytest>=7.0",
    "onnxruntime-gpu>=1.17.0",
    "torch-tensorrt>=2.0.0",
]

[tool.uv.sources]
torch-ema = { git = "https://github.com/fadel/pytorch_ema", rev = "d661859ad4206217afb274be709837d19218b3f8" }

[tool.uv]
environments = ["sys_platform == 'linux' and platform_machine == 'x86_64' and python_full_version < '3.13'"]
